Revert "Merge pull request #18 from sganslandt/prevent-stackoverflow"
That fix was a very weird special case that we shouldn't have to care
about. Making sure we don't rethrow exceptions in general, and making
sure that we don't infinitely recurse for all kinds of objects is a
better general solution.

This reverts commit 2df27a8d3685ab9c44c1be583d45176a2f8d8c3b, reversing
changes made to 916f83d4148584a44d346859f00d8b0693b2d340.